requirements.txt
flask==3.0.2
torch==2.2.1
torchvision==0.17.1
Pillow==10.2.0
numpy==1.26.4
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
opencv-python>=4.5.0
albumentations>=1.0.0
tqdm>=4.62.0
timm>=0.6.0
seaborn>=0.11.0
flask-cors>=3.0.0
gunicorn>=20.0.0 